interactive react component self contained, simplify activating interactives
