+++ packages/client-core/src/common/components/TouchGamepad/TouchGamepadProps.tsx
@@ -1,5 +1,4 @@
 export type TouchGamepadProps = {
-  hovered?: boolean | false
   layout?: string
   onBoardingStep?: number | null
 }
