+++ packages/engine/src/character/CharacterInputSchema.ts
@@ -1,5 +1,4 @@
 import { FollowCameraComponent } from '../camera/components/FollowCameraComponent'
-import { Behavior } from '../common/interfaces/Behavior'
 import { Entity } from '../ecs/classes/Entity'
 import { getComponent } from '../ecs/functions/EntityFunctions'
 import { Input } from '../input/components/Input'
@@ -9,7 +8,7 @@ import { SkinnedMesh } from 'three/src/objects/SkinnedMesh'
 import { CameraModes } from '../camera/types/CameraModes'
 import { LifecycleValue } from '../common/enums/LifecycleValue'
 import { GamepadAxis, XRAxes } from '../input/enums/InputEnums'
-import { getMutableComponent, hasComponent } from '../ecs/functions/EntityFunctions'
+import { getMutableComponent } from '../ecs/functions/EntityFunctions'
 import { CameraInput, GamepadButtons, MouseInput, TouchInputs } from '../input/enums/InputEnums'
 import { InputType } from '../input/enums/InputType'
 import { InputBehaviorType, InputSchema } from '../input/interfaces/InputSchema'
@@ -17,19 +16,15 @@ import { InputAlias } from '../input/types/InputAlias'
 import { Interactable } from '../interaction/components/Interactable'
 import { Interactor } from '../interaction/components/Interactor'
 import { Object3DComponent } from '../scene/components/Object3DComponent'
-import { interactOnServer } from '../interaction/systems/InteractiveSystem'
 import { CharacterComponent } from './components/CharacterComponent'
-import { isClient } from '../common/functions/isClient'
-import { VehicleComponent } from '../vehicle/components/VehicleComponent'
 import { isMobile } from '../common/functions/isMobile'
-import { EquipperComponent } from '../interaction/components/EquipperComponent'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { XRUserSettings, XR_ROTATION_MODE } from '../xr/types/XRUserSettings'
-import { BinaryValue } from '../common/enums/BinaryValue'
 import { ParityValue } from '../common/enums/ParityValue'
-import { XRInputSourceComponent } from './components/XRInputSourceComponent'
 import { InputValue } from '../input/interfaces/InputValue'
 import { NumericalType } from '../common/types/NumericalTypes'
+import { EngineEvents } from '../ecs/classes/EngineEvents'
+import { InteractiveSystem } from '../interaction/systems/InteractiveSystem'
 
 const getParityFromInputValue = (key: InputAlias): ParityValue => {
   switch (key) {
@@ -51,61 +46,16 @@ const getParityFromInputValue = (key: InputAlias): ParityValue => {
 
 const interact = (entity: Entity, inputKey: InputAlias, inputValue: InputValue<NumericalType>, delta: number): void => {
   if (inputValue.lifecycleState !== LifecycleValue.STARTED) return
-
   const parityValue = getParityFromInputValue(inputKey)
 
-  if (!isClient) {
-    //TODO: all this function needs to re-think
-    interactOnServer(entity, parityValue) //TODO: figure out all this cases
-    return
-  }
+  const interactor = getComponent(entity, Interactor)
+  if (!interactor) return
+  const interactive = getComponent(interactor.focusedInteractive, Interactable)
 
-  const equipperComponent = getComponent(entity, EquipperComponent)
-  if (equipperComponent) {
-    return
-  }
-
-  if (!hasComponent(entity, Interactor)) {
-    console.error('Attempted to call interact behavior, but actor does not have Interactor component')
-    return
-  }
-
-  const { focusedInteractive: focusedEntity } = getComponent(entity, Interactor)
-  const input = getComponent(entity, Input)
-  const mouseScreenPosition = input.data.get(BaseInput.SCREENXY)
-
-  // TODO this might be for mobile controls, but breaks non mobile interact
-  // if (mouseScreenPosition && args.phase === LifecycleValue.STARTED ){
-  //   startedPosition.set(entity,mouseScreenPosition.value);
-  //   return;
-  // }
-
-  if (!focusedEntity) {
-    console.log('no focused entity')
-    // no available interactive object is focused right now
-    return
-  }
-
-  if (!hasComponent(focusedEntity, Interactable)) {
-    console.error('Attempted to call interact behavior, but target does not have Interactive component')
-    return
-  }
-
-  const interactive = getComponent(focusedEntity, Interactable)
-  const intPosition = getComponent(focusedEntity, TransformComponent).position
-
-  // TODO: use the aabb of the object instead of the transform to get the accurate position
-  // if (getInteractiveIsInReachDistance(entity, intPosition, args.side)) {
-  if (interactive && typeof interactive.onInteraction === 'function') {
-    if (!hasComponent(focusedEntity, VehicleComponent)) {
-      interactive.onInteraction(entity, { side: parityValue }, delta, focusedEntity)
-    } else {
-      console.log('Interaction with cars must work only from server')
-    }
-  } else {
-    console.warn('onInteraction is not a function')
-  }
-  // }
+  EngineEvents.instance.dispatchEvent({
+    type: InteractiveSystem.EVENTS.OBJECT_ACTIVATION,
+    ...interactive.data
+  })
 }
 
 /**
