+++ packages/engine/src/interaction/systems/InteractiveSystem.ts
@@ -27,7 +27,6 @@ import {
   hasComponent,
   removeComponent
 } from '../../ecs/functions/EntityFunctions'
-import { SystemUpdateType } from '../../ecs/functions/SystemUpdateType'
 import { LocalInputReceiver } from '../../input/components/LocalInputReceiver'
 import { NetworkObject } from '../../networking/components/NetworkObject'
 import { RigidBodyComponent } from '../../physics/components/RigidBody'
@@ -41,9 +40,6 @@ import { Interactable } from '../components/Interactable'
 import { InteractiveFocused } from '../components/InteractiveFocused'
 import { Interactor } from '../components/Interactor'
 import { SubFocused } from '../components/SubFocused'
-import { InteractBehaviorArguments } from '../types/InteractionTypes'
-import { HasHadInteraction } from '../../game/actions/HasHadInteraction'
-import { addActionComponent } from '../../game/functions/functionsActions'
 import { EquipperComponent } from '../components/EquipperComponent'
 import { EquippedStateUpdateSchema } from '../enums/EquippedEnums'
 import { ColliderComponent } from '../../physics/components/ColliderComponent'
@@ -51,88 +47,17 @@ import { NetworkObjectUpdateType } from '../../networking/templates/NetworkObjec
 import { sendClientObjectUpdate } from '../../networking/functions/sendClientObjectUpdate'
 import { BodyType } from 'three-physx'
 import { BinaryValue } from '../../common/enums/BinaryValue'
-import { ParityValue } from '../../common/enums/ParityValue'
-import {
-  getInteractiveIsInReachDistance,
-  interactiveReachDistance
-} from '../../character/functions/getInteractiveIsInReachDistance'
+import { getInteractiveIsInReachDistance } from '../../character/functions/getInteractiveIsInReachDistance'
 import { getHandTransform } from '../../xr/functions/WebXRFunctions'
 import { unequipEntity } from '../functions/equippableFunctions'
 import { Network } from '../../networking/classes/Network'
 import { FontManager } from '../../xrui/classes/FontManager'
-import { isEntityLocalClient } from '../../networking/functions/isEntityLocalClient'
-import { Sphere } from 'three'
-import { TweenComponent } from '../../transform/components/TweenComponent'
-import { Tween } from '@tweenjs/tween.js'
 import { hideInteractText, showInteractText } from '../functions/interactText'
 import { CameraComponent } from '../../camera/components/CameraComponent'
 import { CameraSystem } from '../../camera/systems/CameraSystem'
+import { interactBoxRaycast } from '../functions/interactBoxRaycast'
 
 const vector3 = new Vector3()
-// but is works on client too, i will config out this
-export const interactOnServer: Behavior = (entity: Entity, parityValue, delta): void => {
-  //console.warn('Behavior: interact , networkId ='+getComponent(entity, NetworkObject).networkId);
-
-  const equipperComponent = getComponent(entity, EquipperComponent)
-  if (equipperComponent) {
-    unequipEntity(entity)
-    return
-  }
-
-  let focusedArrays = []
-  for (let i = 0; i < Engine.entities.length; i++) {
-    const isEntityInteractable = Engine.entities[i]
-    if (hasComponent(isEntityInteractable, Interactable)) {
-      const interactive = getComponent(isEntityInteractable, Interactable)
-      const intPosition = getComponent(isEntityInteractable, TransformComponent).position
-      const intRotation = getComponent(isEntityInteractable, TransformComponent).rotation
-      const position = getComponent(entity, TransformComponent).position
-
-      if (interactive.interactionPartsPosition.length > 0) {
-        interactive.interactionPartsPosition.forEach((v, i) => {
-          const partPosition = vector3.set(v[0], v[1], v[2]).applyQuaternion(intRotation).add(intPosition)
-          if (getInteractiveIsInReachDistance(entity, intPosition, parityValue)) {
-            focusedArrays.push([isEntityInteractable, position.distanceTo(partPosition), i])
-          }
-        })
-      } else {
-        if (getInteractiveIsInReachDistance(entity, intPosition, parityValue)) {
-          if (typeof interactive.onInteractionCheck === 'function') {
-            if (interactive.onInteractionCheck(entity, isEntityInteractable, null)) {
-              focusedArrays.push([isEntityInteractable, position.distanceTo(intPosition), null])
-            }
-          } else {
-            focusedArrays.push([isEntityInteractable, position.distanceTo(intPosition), null])
-          }
-        }
-      }
-    }
-  }
-
-  focusedArrays = focusedArrays.sort((a: any, b: any) => a[1] - b[1])
-  if (focusedArrays.length < 1) return
-
-  const interactable = getComponent(focusedArrays[0][0], Interactable)
-  const interactionCheck = interactable.onInteractionCheck(entity, focusedArrays[0][0], focusedArrays[0][2])
-
-  if (interactable.data.interactionType === 'gameobject') {
-    addActionComponent(focusedArrays[0][0], HasHadInteraction, {
-      args: { side: parityValue },
-      entityNetworkId: getComponent(entity, NetworkObject).networkId
-    })
-    return
-  }
-  // Not Game Object
-  if (interactionCheck) {
-    //  console.warn('start with networkId: '+getComponent(focusedArrays[0][0], NetworkObject).networkId+' seat: '+focusedArrays[0][2]);
-    interactable.onInteraction(
-      entity,
-      { side: parityValue, currentFocusedPart: focusedArrays[0][2] },
-      delta,
-      focusedArrays[0][0]
-    )
-  }
-}
 
 export const subFocused: Behavior = (entity: Entity, args, delta: number): void => {
   if (!hasComponent(entity, Interactable)) {
@@ -151,124 +76,11 @@ const interactFocused: Behavior = (entity: Entity, args, delta: number): void =>
   }
 
   const focused = hasComponent(entity, InteractiveFocused)
-  const subFocused = hasComponent(entity, SubFocused)
 
   const interactive = getComponent(entity, Interactable)
 
-  if (interactive && typeof interactive.onInteractionFocused === 'function') {
-    const entityFocuser = focused ? getComponent(entity, InteractiveFocused).interacts : null
-    interactive.onInteractionFocused(entityFocuser, { focused }, delta, entity)
-  }
-}
-
-const mat4 = new Matrix4()
-const projectionMatrix = new Matrix4().makePerspective(
-  -0.1, // x1
-  0.1, // x2
-  -0.1, // y1
-  0.1, // y2
-  0.1, // near
-  2 // far
-)
-const frustum = new Frustum()
-const vec3 = new Vector3()
-
-type RaycastResult = [Entity, boolean, number?, number?]
-
-/**
- * Checks if entity can interact with any of entities listed in 'interactive' array, checking distance, guards and raycast
- * @param entity
- * @param interactive
- * @param delta
- */
-
-const interactBoxRaycast: Behavior = (
-  entity: Entity,
-  { raycastList }: InteractBehaviorArguments,
-  delta: number
-): void => {
-  const interacts = getMutableComponent(entity, Interactor)
-  if (!isEntityLocalClient(entity)) {
-    interacts.subFocusedArray = []
-    interacts.focusedInteractive = null
-    return
-  }
-
-  const transform = getComponent(entity, TransformComponent)
-  const actor = getComponent(entity, CharacterComponent)
-
-  if (!raycastList.length) {
-    return
-  }
-
-  actor.frustumCamera.updateMatrixWorld()
-  actor.frustumCamera.matrixWorldInverse.copy(actor.frustumCamera.matrixWorld).invert()
-
-  mat4.multiplyMatrices(projectionMatrix, actor.frustumCamera.matrixWorldInverse)
-  frustum.setFromProjectionMatrix(mat4)
-
-  const subFocusedArray = raycastList
-    .map((entityIn): RaycastResult => {
-      const boundingBox = getComponent(entityIn, BoundingBoxComponent)
-      const interactive = getComponent(entityIn, Interactable)
-      if (boundingBox.boxArray.length) {
-        // TO DO: static group object
-
-        if (boundingBox.dynamic) {
-          const arr = boundingBox.boxArray
-            .map((object3D, index): RaycastResult => {
-              if (interactive.onInteractionCheck(entity, entityIn, index)) {
-                const aabb = new Box3()
-                aabb.setFromObject(object3D)
-                return [entityIn, frustum.intersectsBox(aabb), aabb.distanceToPoint(transform.position), index]
-              }
-              return [entityIn, false, null, index]
-            })
-            .filter((value) => value[1])
-            .sort((a: any, b: any) => a[2] - b[2])
-
-          if (arr.length) {
-            return arr[0]
-          } else {
-            return [null, false]
-          }
-        }
-      } else {
-        if (boundingBox.dynamic) {
-          const object3D = getComponent(entityIn, Object3DComponent)
-          const aabb = new Box3()
-          aabb.copy(boundingBox.box)
-          aabb.applyMatrix4(object3D.value.matrixWorld)
-          return [entityIn, frustum.intersectsBox(aabb), aabb.distanceToPoint(transform.position)]
-        } else {
-          return [entityIn, frustum.intersectsBox(boundingBox.box), boundingBox.box.distanceToPoint(transform.position)]
-        }
-      }
-    })
-    .filter((value) => value[1])
-
-  if (!subFocusedArray.length) {
-    interacts.subFocusedArray = []
-    interacts.focusedInteractive = null
-    return
-  }
-
-  interacts.subFocusedArray = subFocusedArray.map((v: any) => [getComponent(v[0], Object3DComponent).value, v[3]])
-
-  const [entityInteractable, doesIntersectFrustrum, distanceToPlayer] = subFocusedArray.sort(
-    (a: any, b: any) => a[2] - b[2]
-  )[0]
-
-  const interactable = getComponent(entityInteractable, Interactable)
-  const distance =
-    typeof interactable.data.interactionDistance !== 'undefined'
-      ? interactable.data.interactionDistance
-      : interactiveReachDistance
-
-  const resultIsCloseEnough = distanceToPlayer < distance
-  if (resultIsCloseEnough) {
-    interacts.focusedInteractive = entityInteractable
-  }
+  const entityFocuser = focused ? getComponent(entity, InteractiveFocused).interacts : null
+  // on interactive focused
 }
 
 const upVec = new Vector3(0, 1, 0)
@@ -338,7 +150,7 @@ export class InteractiveSystem extends System {
     if (isClient) {
       this.queryResults.interactors?.all.forEach((entity) => {
         if (this.queryResults.interactive?.all.length) {
-          interactBoxRaycast(entity, { raycastList: this.queryResults.boundingBox.all })
+          interactBoxRaycast(entity, this.queryResults.boundingBox.all)
           const interacts = getComponent(entity, Interactor)
           if (interacts.focusedInteractive) {
             this.newFocused.add(interacts.focusedInteractive)
@@ -384,23 +196,18 @@ export class InteractiveSystem extends System {
         object3D.rotation.setFromQuaternion(transform.rotation)
         if (!calcBoundingBox.dynamic) object3D.updateMatrixWorld()
 
-        if (interactive.interactionParts.length) {
-          const arr = interactive.interactionParts.map((name) => object3D.children[0].getObjectByName(name))
-          calcBoundingBox.boxArray = arr
-        } else {
-          object3D.traverse((obj3d: Mesh) => {
-            if (obj3d instanceof Mesh) {
-              if (!obj3d.geometry.boundingBox) obj3d.geometry.computeBoundingBox()
-              const aabb = new Box3().copy(obj3d.geometry.boundingBox)
-              if (!calcBoundingBox.dynamic) aabb.applyMatrix4(obj3d.matrixWorld)
-              if (!calcBoundingBox.box) {
-                calcBoundingBox.box = aabb
-              } else {
-                calcBoundingBox.box.union(aabb)
-              }
+        object3D.traverse((obj3d: Mesh) => {
+          if (obj3d instanceof Mesh) {
+            if (!obj3d.geometry.boundingBox) obj3d.geometry.computeBoundingBox()
+            const aabb = new Box3().copy(obj3d.geometry.boundingBox)
+            if (!calcBoundingBox.dynamic) aabb.applyMatrix4(obj3d.matrixWorld)
+            if (!calcBoundingBox.box) {
+              calcBoundingBox.box = aabb
+            } else {
+              calcBoundingBox.box.union(aabb)
             }
-          })
-        }
+          }
+        })
       }
 
       // removal is the first because the hint must first be deleted, and then a new one appears
